refactor: 바텀 네비게이션 로직 병합
