+++ lib/presentation/pages/main/main_event.dart
@@ -1,5 +1,4 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:techtalk/core/constants/main_navigation_tab.enum.dart';
 import 'package:techtalk/presentation/providers/main_bottom_navigation_provider.dart';
 
 abstract interface class _MainEvent {
